+++ packages/fighting-design/date-picker/index.ts
@@ -4,6 +4,7 @@ import { install } from '../_utils'
 
 export const FDatePicker = install(DatePicker)
 
+/** date-picker 组件实例类型 */
 export type DatePickerInstance = InstanceType<typeof DatePicker>
 
 export * from './src/interface.d'