+++ packages/fighting-design/date-picker/src/props.ts
@@ -6,8 +6,8 @@ import type { FightingSize } from '../../_interface'
 export const Props = {
   /** 绑定值 */
   date: setStringProp(),
-  /** 是否只读 */
-  readonly: setBooleanProp(),
+  /** 是否禁用 */
+  disabled: setBooleanProp(),
   /** 是否可清除 */
   clear: setBooleanProp(),
   /**
@@ -22,7 +22,10 @@ export const Props = {
   /** 时间格式 */
   format: setStringProp<string>('YYYY/MM/DD'),
   /** 占位符 */
-  placeholder: setStringProp()
+  placeholder: setStringProp(),
+  /** 不满 10 的是否前面加 0 */
+  addZero: setBooleanProp()
 } as const
 
+/** date-picker 组件 props 类型 */
 export type DatePickerProps = ExtractPropTypes<typeof Props>