+++ packages/fighting-design/_utils/compatible/index.ts
@@ -16,9 +16,9 @@ import { warning } from '..'
  * @returns { Array } 给定对象自身可枚举属性的键值对数组
  */
 export const objectEntries = <T extends object>(obj: T): [keyof T, T[keyof T]][] => {
-  /** 
+  /**
    * 如果不是对象类型返回空数组
-   * 
+   *
    * 并且提示警告
    */
   if (typeof obj !== 'object' || obj === null) {
@@ -45,23 +45,22 @@ export const objectEntries = <T extends object>(obj: T): [keyof T, T[keyof T]][]
   for (const key in obj) {
     /**
      * 必须对象自身属性中是否具有指定的属性
-     * 
+     *
      * 对象可能具有隐藏内置函数的属性 Object.prototype，可能会导致意外行为或拒绝服务安全漏洞
-     * 
+     *
      * 所以这里没有直接使用 obj.hasOwnProperty(key) 而是改用 Object.prototype.hasOwnProperty.call(obj, key)
-     * 
+     *
      * @see no-prototype-builtins https://eslint.org/docs/latest/rules/no-prototype-builtins
-     * 
+     *
      * 方法详情参考
-     * 
+     *
      * @see Object.prototype.hasOwnProperty() https://developer.mozilla.org/zh-CN/docs/Web/JavaScript/Reference/Global_Objects/Object/hasOwnProperty
-     * 
+     *
      * 兼容性参考
-     * 
+     *
      * @see hasOwnProperty https://caniuse.com/?search=hasOwnProperty
      */
     if (Object.prototype.hasOwnProperty.call(obj, key)) {
-      console.log(1)
       /**
        * key 为第一项，value 为第二项组成的组数
        *