+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - `f-slider` 组件新增 `background` `active-color` `width` 配置项
 - `f-card` 组件新增 `border-color` `background` `title-color` `title-bold` `close` `on-close` 配置项
 - `f-input` 组件新增 `autocomplete` 配置项
+- `f-input` 组件新增 `before` 插槽
+
+- 修复 `f-date-picker` 清空时的警告提示
 
 ## 0.22.1 (2023-02-07)